refactor(api): extract the system-update subsystem into src/system/ (pure move)
~1,200 lines of the update-check / self-update / snapshot family move out of
src/api_server.py:

* src/system/update_service.py — the subsystem as a service: Docker Hub +
  GitHub version comparison (newest-of-both-sources, #1430, verbatim), the
  fiestaupdater sidecar client, pre-update settings snapshots with retention
  pruning, and the .system-update.json state machine with its RLock +
  write_json_atomic semantics (#1745) preserved exactly. Gains one new
  function, run_system_update_check_if_due(), the tick body of the hourly
  scheduled check; the loop shell stays in api_server's lifespan.
* src/system/models.py — the nine request/response models, verbatim.
* src/system/routes.py — the eight handlers (/version, /system/update-check,
  /system/update/status, POST /system/update, /system/update/rollback,
  /system/update/auto, /system/restart, /system/shutdown), verbatim, thin on
  the service.

Patch-seam disposition (the #1756/#1757 pattern):

* api_server re-imports every moved name, so patch("src.api_server.<name>")
  targets keep working; the extracted handlers import those names back
  through src.api_server at call time.
* The two path-override test seams (SYSTEM_UPDATE_STATE_FILE /
  SETTINGS_SNAPSHOT_DIR) stay *defined* on api_server; the service's
  _system_update_state_file()/_settings_snapshot_dir() read them back
  through api_server at call time, so monkeypatching them there still
  steers the service. No test fixture changed.
* The boot-time post-upgrade family (_build_post_upgrade_restore_set,
  _auto_restore_post_upgrade_regression, _log_config_boot_snapshot,
  _detect_post_upgrade_regression) stays in api_server: it runs from the
  lifespan and the suite monkeypatches api_server._resolve_snapshot_name /
  get_config_manager around direct api_server.<name>() calls.
* _auto_apply_plugin_updates stays: it belongs to the plugin update loop,
  not the system-update subsystem.

The new src/system files inherit api_server's lint-ignore set in
pyproject.toml, same as the #1756 routers (cleanup with #1766).

Zero-regression evidence: tests/test_system_endpoints.py passes unmodified,
route-inventory golden unchanged, response-shape goldens zero-diff; full
suite (-m "not integration" -n auto) 5859 passed -> 5860 passed (the one
addition is the new system golden), same single pre-existing environmental
failure (test_check_image_formats git-worktree issue) before and after.

Part of #1849; closes #1758.
